moved createpost from pages to components
